Move non-CI presets to CMakeUserPresets.json
A template CMakeUserPresets.json file has been added to the "cmake/"
directory that can be optionally copied and used to provide an extended
set of presets that we don't necessarily CI against.

This cleans up the preset list for supported/default builds.

The presets that have been relocated are the architecture-specific
Windows ones (windows-amd64 and windows-x86) as well as the debug
build presets for all target systems.