Rename project from tainted-lua to elune
Quite frankly I'm ashamed I didn't think of this name back at the
inception of the project.

As the scope of this is increasing over time and we're not _just_
implementing taint but also profiling and base library changes, it
makes sense for the project name and description to reflect that.

The README still requires some (large) changes to accomodate, but
we'll now market ourselves as elune.