Generate luaconf.h from CMake cache settings
This fairly substantially rewrites the configuration header to instead
be generated at configure-time per variables in the CMake cache and
from detected system APIs.

This makes configuring a few aspects a bit easier - such as the Lua
library search paths - and also makes consuming the build a bit easier
as all of the necessary state is encoded in the header, thus no
additional defines should need adding if building or linking against
the project externally.