Initial work on standard library split
The idea behind these changes is to offer two standard libraries within
the Lua core; the default one with the full API set and a reduced
WoW-compatible subset.

In most cases API functions will be shared between both of these
libraries - however for any potentially incompatible APIs there'll
be differences between the API sets; such as loadstring won't taint
in the "full" API set but will in the "reduced" one.