Add 'base' library to WoW standard libraries
This doesn't bring in any major changes aside from fixing the pain point
of loadstring being insecure under the "normal" standard library.

One slight deviant from reference here though is that we do provide an
implementation of 'print' in the WoW-compatible standard library, which
isn't necessarily accurate. However - with it we also implement the
full print handler framework (and tostringall) with it, all as Lua
functions to deal with any edge cases where code might be expecting
these to be Lua functions.