Flatten directory hierarchy one level
Remove the 'src/' directory and bring all its subfolders to the top
of the repository as it isn't of much use the way it is.

The 'include/' directory has been relocated to be a subdirectory
of the 'liblua/' folder.

Note that - for now - 'lsyslib.h' will also be installed along with
other public headers. This will likely change soon.