Relocate abstraction APIs to lauxlib.h
This removes the 'lsyslib.h' header thus making it no longer part of
the installed header set. The functions have had their 'luaL_' prefix
returned to them and can be considered "stable" parts of the auxillary
library API now.

Note that 'luaI_clockrate' and 'luaI_clocktime' are _not_ exported APIs;
these are used internally by 'ldebug.c' and won't be exposed as it'd
be a bit confusing to have both those and the normal 'lua_clock*' APIs
present.