Remove 'LUA_USE_DLOPEN' option
This doesn't seem to work on macOS due to the dependency check on the
CMAKE_DL_LIBS variable - which apparently is empty on that platform;
presumbly dlopen and friends are in the set of standard libraries
and thus the list is empty?

Either way as it's not too likely someone will want to explicitly
disable linking against the dynamic loader implementation on a platform
we'll remove the option to do so and will instead gate the use of
dlopen and friends on 'LUA_USE_POSIX'.