Add a dirty cast to silence a macOS build warning
The 'rl_readline_name' variable is officially declared as a 'char*' so
Clang rightly diagnoses that we're assigning a 'const char*' to it
and complains.

In reality, other projects _also_ just use string literals and do
the same cast - so it's probably safe to assume no one's going to
actually mutate the pointed-to string.