Format signed values as longs correctly
Lua formats the %d and %i specifiers as 'long' by default, however we
were inappropriately casting numeric inputs to 'int'. On Windows this
results in no difference, but on Unix platforms this is incorrect and
leads to a situation where we'd lose data unnecessarily.

A new lua_tolong API has been added to obtain and convert a numeric
value on the stack to a long. Due to a fun reference compatibility
quirk, the game will format %u as 64-bit values on Mac (indicating
that it is indeed using 'long', not 'int') _but_ will unexpectedly
reject values outside the representable range of 'int' for %d/%i.

To replicate this quirk, when overflow checks are enabled we'll
ensure that errors on conversion to 'long' are raised if the value
to be converted remains outside the range of 'int'. Not the cleanest
or the best, but it is accurate.

Note that to simplify matters, the LUA_USE_LONGLONG define has been
removed. As it's clear the reference environment doesn't use it, we
don't need it either.